% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_cor_funs.R
\name{cor_fmt}
\alias{cor_fmt}
\title{importFrom("stats", "p.adjust")}
\usage{
cor_fmt(
  df,
  cor_method = c("spearman"),
  col_list = NULL,
  mc_method = NULL,
  group = NULL
)
}
\arguments{
\item{df}{data frame or tibble of data}

\item{cor_method}{choose from "spearman" or "pearson"}

\item{col_list}{optional list of columns to compare with each other (if not all columns against all other columns)}

\item{mc_method}{mutliple comparison method; see ?p.adjust for more details}

\item{group}{column to split comparisons by; must contain two unique values}
}
\value{
tibble with ready to format correlation data
}
\description{
importFrom("stats", "p.adjust")
}
