% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_2groups.R
\name{cor_fmt_2g}
\alias{cor_fmt_2g}
\title{Format correlations for two different groups for use in a heat map}
\usage{
cor_fmt_2g(df, cor_method, group)
}
\arguments{
\item{df}{A data frame or tibble containing only the columns you want to calculate the correlations of, and a group identifier column.}

\item{group}{A string containing the name of a column which identifies the two groups you want correlations computed.}
}
\value{
a tidy tibble ready for plotting
}
\description{
Format correlations for two different groups for use in a heat map
}
\examples{
format_cor_2groups(mtcars, "am")

}
